Add clippy check to sandbox CI
Wires `cargo clippy --workspace --all-targets -- -D warnings` into the
sandbox native (Rust) workspace CI and clears the workspace to zero
warnings so the gate passes.

Workflow (`sandbox-check.yml`):
- add the `clippy` component to the Rust toolchain
- add a "Run Rust clippy" step. It runs from the workspace directory
  (not `--manifest-path` from the repo root) so the crate's
  `.cargo/config.toml` — which sets `--cfg tokio_unstable`, required by
  the tokio `RuntimeMetrics` calls in stats.rs / merge/metrics.rs — is
  picked up. cargo only reads `.cargo/config.toml` relative to the CWD.

Code changes (bulk is mechanical clippy cleanup):
- fix deny-by-default lints that were aborting the lint build
  (not_unsafe_ptr_arg_deref on FFI entry points, never_loop,
  absurd_extreme_comparisons, approx_constant, mut_from_ref)
- migrate deprecated parquet/arrow APIs (with_page_index ->
  with_page_index_policy, set_max_row_group_size ->
  set_max_row_group_row_count)
- repair stale benchmarks that no longer matched current API signatures
- the remaining diff is unused-import/mut removal, doc formatting,
  and idiomatic simplifications

Two deprecated calls are intentionally kept behind `#[allow(deprecated)]`:
`PartitionedFile::with_extensions` in parquet_bridge.rs and
shard_table_provider.rs. The `with_extension` replacement keys the
extension by concrete type, but DataFusion's Parquet opener retrieves the
`ParquetAccessPlan` via the legacy `insert_dyn` keying — switching drops
row selection and returns wrong rows. A comment documents the deferral.

clippy (-D warnings), fmt --check, and the test suite all pass.

+++ sandbox/plugins/analytics-backend-datafusion/rust/src/udf/extract.rs
@@ -12,8 +12,6 @@
 
 use std::sync::Arc;
 
-use super::udf_identity;
-
 use chrono::{DateTime, Datelike, NaiveTime, TimeZone, Timelike, Utc};
 use datafusion::arrow::array::{
     Array, ArrayRef, AsArray, Int64Builder, StringArray, Time32MillisecondArray, Time32SecondArray,
@@ -284,7 +282,7 @@ fn extract_for_unit(unit: &str, dt: DateTime<Utc>) -> Option<i64> {
         "DAY" => Some(dd),
         "WEEK" => Some(dt.iso_week().week() as i64),
         "MONTH" => Some(mo),
-        "QUARTER" => Some(((mo - 1) / 3 + 1) as i64),
+        "QUARTER" => Some((mo - 1) / 3 + 1),
         "YEAR" => Some(yy),
         "DOW" => Some(dt.weekday().number_from_monday() as i64),
         "DOY" => Some(dt.ordinal() as i64),
